refactor MapAreaDetailsScreen
